+++ src/types/globalTypes.ts
@@ -89,6 +89,7 @@ export enum AttributeInputTypeEnum {
   DROPDOWN = "DROPDOWN",
   FILE = "FILE",
   MULTISELECT = "MULTISELECT",
+  NUMERIC = "NUMERIC",
   REFERENCE = "REFERENCE",
   RICH_TEXT = "RICH_TEXT",
 }
@@ -528,6 +529,39 @@ export enum LanguageCodeEnum {
   VI = "VI",
   ZH_HANS = "ZH_HANS",
   ZH_HANT = "ZH_HANT",
+}
+
+export enum MeasurementUnitsEnum {
+  ACRE_FT = "ACRE_FT",
+  ACRE_IN = "ACRE_IN",
+  CM = "CM",
+  CUBIC_CENTIMETER = "CUBIC_CENTIMETER",
+  CUBIC_DECIMETER = "CUBIC_DECIMETER",
+  CUBIC_FOOT = "CUBIC_FOOT",
+  CUBIC_INCH = "CUBIC_INCH",
+  CUBIC_METER = "CUBIC_METER",
+  CUBIC_MILLIMETER = "CUBIC_MILLIMETER",
+  CUBIC_YARD = "CUBIC_YARD",
+  FL_OZ = "FL_OZ",
+  FT = "FT",
+  G = "G",
+  INCH = "INCH",
+  KG = "KG",
+  KM = "KM",
+  LB = "LB",
+  LITER = "LITER",
+  M = "M",
+  OZ = "OZ",
+  PINT = "PINT",
+  QT = "QT",
+  SQ_CM = "SQ_CM",
+  SQ_FT = "SQ_FT",
+  SQ_INCH = "SQ_INCH",
+  SQ_KM = "SQ_KM",
+  SQ_M = "SQ_M",
+  SQ_YD = "SQ_YD",
+  TONNE = "TONNE",
+  YD = "YD",
 }
 
 export enum MenuErrorCode {
@@ -1025,6 +1059,7 @@ export enum WeightUnitsEnum {
   KG = "KG",
   LB = "LB",
   OZ = "OZ",
+  TONNE = "TONNE",
 }
 
 export interface AddressInput {
@@ -1075,6 +1110,7 @@ export interface AttributeCreateInput {
   name: string;
   slug?: string | null;
   type: AttributeTypeEnum;
+  unit?: MeasurementUnitsEnum | null;
   values?: (AttributeValueCreateInput | null)[] | null;
   valueRequired?: boolean | null;
   isVariantOnly?: boolean | null;
@@ -1104,6 +1140,7 @@ export interface AttributeFilterInput {
 export interface AttributeInput {
   slug: string;
   values?: (string | null)[] | null;
+  valuesRange?: IntRangeInput | null;
 }
 
 export interface AttributeSortingInput {
@@ -1114,6 +1151,7 @@ export interface AttributeSortingInput {
 export interface AttributeUpdateInput {
   name?: string | null;
   slug?: string | null;
+  unit?: MeasurementUnitsEnum | null;
   removeValues?: (string | null)[] | null;
   addValues?: (AttributeValueCreateInput | null)[] | null;
   valueRequired?: boolean | null;